ansible/roles/flask/files/requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
Werkzeug==2.3.7
